core/install-arch-linux: update notes on AUR, drop sudo in yaourt invocation
The use of helpers for installing packages from AUR is a sensitive topic in
ArchLinux community. Officially none of the AUR helpers is blessed and the users
are 'expected' to inspect the build files themselves.

Rephrase the installation instructions for Arch so that we try to keep the
status quo.

Drop `sudo` from yaourt invocation as yaourt will run it itself just before
running pacman to install the built packages.. The way the helper is written,
it's best to avoid running it under sudo.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>